Add retry policy for TestRailClient
Added a retry policy to the TestRailClient for handling transient errors. RetryPolicyBuilder class is introduced to create a retry policy. The policy can handle exceptions like RequestTimeout, BadGateway, ServiceUnavailable and GatewayTimeout by executing retries with exponential backoff. The Policy package reference was included to the project to implement the retry policies.